cleanup file structure
